@@ -31,11 +31,11 @@
     }
   ],
   "require": {
-    "php": ">=5.3.3"
+    "php": "^7"
   },
   "require-dev": {
     "jakub-onderka/php-parallel-lint": "^0.9",
-    "phpunit/phpunit": "^4.0.0"
+    "phpunit/phpunit": "^5"
   },
   "suggest": {
     "ext-gmp": "Large integer support for 32-bit platforms."
