@@ -7,9 +7,7 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,23 +15,23 @@
  * limitations under the License.
  */
 
-require_once('test_helper.php');
+require_once 'test_helper.php';
 
 /**
  * Class IODatumReaderTest
  */
-class IODatumReaderTest extends PHPUnit_Framework_TestCase
+class IODatumReaderTest extends PHPUnit\Framework\TestCase
 {
 
-  public function testSchemaMatching()
-  {
-    $writers_schema = <<<JSON
+    public function testSchemaMatching()
+    {
+        $writers_schema = <<<JSON
       { "type": "map",
         "values": "bytes" }
 JSON;
-    $readers_schema = $writers_schema;
-    $this->assertTrue(AvroIODatumReader::schemas_match(
-                        AvroSchema::parse($writers_schema),
-                        AvroSchema::parse($readers_schema)));
-  }
+        $readers_schema = $writers_schema;
+        $this->assertTrue(AvroIODatumReader::schemas_match(
+            AvroSchema::parse($writers_schema),
+            AvroSchema::parse($readers_schema)));
+    }
 }
