@@ -7,9 +7,7 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,67 +15,69 @@
  * limitations under the License.
  */
 
-require_once('test_helper.php');
+require_once 'test_helper.php';
 
 /**
  * Class InterOpTest
  */
-class InterOpTest extends PHPUnit_Framework_TestCase
+class InterOpTest extends PHPUnit\Framework\TestCase
 {
-  var $projection_json;
-  var $projection;
+    public $projection_json;
+    public $projection;
 
-  public function setUp()
-  {
-    $interop_schema_file_name = AVRO_INTEROP_SCHEMA;
-    $this->projection_json = file_get_contents($interop_schema_file_name);
-    $this->projection = AvroSchema::parse($this->projection_json);
-  }
+    public function setUp()
+    {
+        $interop_schema_file_name = AVRO_INTEROP_SCHEMA;
+        $this->projection_json = file_get_contents($interop_schema_file_name);
+        $this->projection = AvroSchema::parse($this->projection_json);
+    }
 
-  /**
-   * @return array
-   */
-  public function file_name_provider()
-  {
-    $data_dir = AVRO_BUILD_DATA_DIR;
-    $data_files = array();
-    if (!($dh = opendir($data_dir)))
-      die("Could not open data dir '$data_dir'\n");
+    /**
+     * @return array
+     */
+    public function file_name_provider()
+    {
+        $data_dir = AVRO_BUILD_DATA_DIR;
+        $data_files = array();
+        if (!($dh = opendir($data_dir)))
+            die("Could not open data dir '$data_dir'\n");
 
-    /* TODO This currently only tries to read files of the form 'language.avro',
-     * but not 'language_deflate.avro' as the PHP implementation is not yet
-     * able to read deflate data files. When deflate support is added, change
-     * this to match *.avro. */
-    while ($file = readdir($dh))
-      if (0 < preg_match('/^[a-z]+\.avro$/', $file))
-        $data_files []= join(DIRECTORY_SEPARATOR, array($data_dir, $file));
-    closedir($dh);
+        /* TODO This currently only tries to read files of the form 'language.avro',
+         * but not 'language_deflate.avro' as the PHP implementation is not yet
+         * able to read deflate data files. When deflate support is added, change
+         * this to match *.avro. */
+        while ($file = readdir($dh))
+            if (0 < preg_match('/^[a-z]+\.avro$/', $file))
+                $data_files [] = implode(DIRECTORY_SEPARATOR, array($data_dir, $file));
+        closedir($dh);
 
-    $ary = array();
-    foreach ($data_files as $df)
-      $ary []= array($df);
-    return $ary;
-  }
+        $ary = array();
+        foreach ($data_files as $df)
+            $ary [] = array($df);
+        return $ary;
+    }
 
-  /**
-   * @dataProvider file_name_provider
-   * @param $file_name
-   * @throws AvroDataIOException
-   */
-  public function test_read($file_name)
-  {
+    /**
+     * @dataProvider file_name_provider
+     *
+     * @param $file_name
+     *
+     * @throws AvroDataIOException
+     */
+    public function test_read($file_name)
+    {
 
-    $dr = AvroDataIO::open_file(
-      $file_name, AvroFile::READ_MODE, $this->projection_json);
+        $dr = AvroDataIO::open_file(
+            $file_name, AvroFile::READ_MODE, $this->projection_json);
 
-    $data = $dr->data();
+        $data = $dr->data();
 
-    $this->assertNotEquals(0, count($data),
-                           sprintf("no data read from %s", $file_name));
+        $this->assertNotEquals(0, count($data),
+            sprintf("no data read from %s", $file_name));
 
-    foreach ($data as $idx => $datum)
-      $this->assertNotNull($datum, sprintf("null datum from %s", $file_name));
+        foreach ($data as $idx => $datum)
+            $this->assertNotNull($datum, sprintf("null datum from %s", $file_name));
 
-  }
+    }
 
 }
